worker_server/requirements.txt
pika
schedule
requests
python-dotenv